Support for CRYSTALS-Dilithium (NIST PQC Round2)
https://pq-crystals.org/dilithium/data/dilithium-specification-round2.pdf

This implementation is based on the spec with the reference
implementation used as, what is in a name, a reference.  It turns out
to be pretty close to the reference implementation.  I did not
simply mimic their choices, but checked them (for instance, whether
coefficients stay bounded.)  I did not find an issue except for
some errors in comments.

If available, AVX2 is used to speed up the NTT.  Our AVX2 NTT and
inverse-NTT implementations are different from the optimized reference
implementation.  I chose to use a simpler "scheduling" for ease of
review.  Our NTT is a tiny bit faster, but our inverse-NTT is a bit slower.

There is still room for improvement: the ExpandA (i.e. Mat.Derive)
and ExpandY (i.e. PolyDeriveUniformLeGamma1) can be optimized with AVX2.

There is a considerable amount of precomputation that can be shared
between the computation of multiple sigantures with the same private
key.  We do this precomputation when unpacking the private key and store
it in the PrivateKey struct.   Same applies for verification of multiple
signatures for a public key.

Performance (coffee lake, 8569U) for Dilithium mode 3 using AVX2.

    BenchmarkSkUnpack-8            	   18980	     61840 ns/op
    BenchmarkPkUnpack-8            	   19743	     60087 ns/op
    BenchmarkVerify-8              	   39998	     30421 ns/op
    BenchmarkSign-8                	    5420	    228631 ns/op
    BenchmarkGenerateKey-8         	   12037	     99672 ns/op
    BenchmarkPublicFromPrivate-8   	  107148	     11512 ns/op

Non-optmized C reference implementation

    BenchmarkKeygen-8   	   10000	    109216 ns/op
    BenchmarkSign-8     	    2206	    545616 ns/op
    BenchmarkVerify-8   	   10000	    108419 ns/op

AVX2 optimized C reference implementation

    BenchmarkKeygen-8   	   36510	     32242 ns/op
    BenchmarkSign-8     	   10000	    117060 ns/op
    BenchmarkVerify-8   	   31605	     34123 ns/op

To make a fair comparisson (due to the precomputations), one needs to
take Sign+SkUnpack and Verify+PkUnpack together.  So that is:

    keygen   99µs (our) vs 109µs (cref) and  32µs (avx2ref)
    signing 290µs (our) vs 546µs (cref) and 117µs (avx2ref)
    verif    60µs (our) vs 108µs (cref) and  34µs (avx2ref)