sign: expose seed of ML-DSA private keys
Alternatively, we could've switched ML-DSA private keys to seeds and expose
the expanded private key via an interface. That'd break backwards
compatibility (not that big of a deal for freshly minted crypto), but worse
it'd be confusing for ecosystems that follow the NIST ecosystem instead of X509.
I'm on the fence. This seemed to be the simpler approach.